--- src/main/java/com/stemm/pubsub/service/post/entity/LikeStatus.java
@@ -1,6 +1,0 @@
-package com.stemm.pubsub.service.post.entity;
-
-public enum LikeStatus {
-    LIKE,
-    DISLIKE
-}
