+++ src/main/java/com/stemm/pubsub/service/post/entity/post/LikeStatus.java
@@ -1,0 +1,6 @@
+package com.stemm.pubsub.service.post.entity.post;
+
+public enum LikeStatus {
+    LIKE,
+    DISLIKE
+}
