+++ src/main/java/com/stemm/pubsub/service/post/entity/post/PostLike.java
@@ -1,4 +1,4 @@
-package com.stemm.pubsub.service.post.entity;
+package com.stemm.pubsub.service.post.entity.post;
 
 import com.stemm.pubsub.common.BaseEntity;
 import com.stemm.pubsub.service.user.entity.User;
@@ -12,33 +12,29 @@ import static jakarta.persistence.GenerationType.IDENTITY;
 import static lombok.AccessLevel.PROTECTED;
 
 @Entity
-@Table(indexes = @Index(name = "idx_post_created_date", columnList = "created_date"))
 @NoArgsConstructor(access = PROTECTED)
 @Getter
-public class Post extends BaseEntity {
+public class PostLike extends BaseEntity {
     @Id
     @GeneratedValue(strategy = IDENTITY)
-    @Column(name = "post_id")
+    @Column(name = "post_like_id")
     private Long id;
+
+    @ManyToOne(fetch = LAZY)
+    @JoinColumn(name = "post_id")
+    private Post post;
 
     @ManyToOne(fetch = LAZY)
     @JoinColumn(name = "user_id")
     private User user;
 
-    @Lob
-    @Column(nullable = false)
-    private String content;
-
-    private String imageUrl;
-
     @Enumerated(STRING)
     @Column(nullable = false)
-    private Visibility visibility;
+    private LikeStatus status;
 
-    public Post(User user, String content, String imageUrl, Visibility visibility) {
+    public PostLike(Post post, User user, LikeStatus status) {
+        this.post = post;
         this.user = user;
-        this.content = content;
-        this.imageUrl = imageUrl;
-        this.visibility = visibility;
+        this.status = status;
     }
 }
