+++ src/main/java/com/stemm/pubsub/service/post/entity/post/Visibility.java
@@ -1,4 +1,4 @@
-package com.stemm.pubsub.service.post.entity;
+package com.stemm.pubsub.service.post.entity.post;
 
 public enum Visibility {
     PUBLIC,
